mm/memcontrol: make memory.oom_guarantee file also available in cgroup-v2
The interface is slightly reworked to be more v2 like:

 - switch from "-1" to "max" for max value following cgroup-v2 style

While on it also:
 - make oom_guarantee (unsigned long) similar to regular cgroup limits,
   as we never set more than PAGE_COUNTER_MAX there anyway
 - set oom_guarantee with xchg(), this ensures that update is atomic

https://virtuozzo.atlassian.net/browse/PSBM-154224

Feature: mm/oom: OOM guarantee feature

+++ include/linux/memcontrol.h
@@ -263,7 +263,7 @@ struct mem_cgroup {
 	struct vmpressure vmpressure;
 
 	unsigned long overdraft;
-	unsigned long long oom_guarantee;
+	unsigned long oom_guarantee;
 
 	/*
 	 * Should the OOM killer kill all belonging tasks, had it kill one?
